passing test, lots of cleanup
