+++ acceptance/tests/fixtures/bases/static-client/deployment.yaml
@@ -18,7 +18,7 @@ spec:
     spec:
       containers:
         - name: static-client
-          image: docker.mirror.hashicorp.services/hashicorp/http-echo:alpine
+          image: docker.mirror.hashicorp.services/buildpack-deps:jammy-curl
           command: [ "/bin/sh", "-c", "--" ]
           args: [ "while true; do sleep 30; done;" ]
           env:
