+++ acceptance/tests/fixtures/bases/v2-multiport-app/psp-rolebinding.yaml
@@ -12,15 +12,3 @@ roleRef:
 subjects:
   - kind: ServiceAccount
     name: multiport
----
-apiVersion: rbac.authorization.k8s.io/v1
-kind: RoleBinding
-metadata:
-  name: multiport-admin
-roleRef:
-  apiGroup: rbac.authorization.k8s.io
-  kind: ClusterRole
-  name: test-psp
-subjects:
-  - kind: ServiceAccount
-    name: multiport-admin
