+++ acceptance/tests/fixtures/bases/v2-multiport-app/secret.yaml
@@ -4,14 +4,6 @@
 apiVersion: v1
 kind: Secret
 metadata:
-  name: multiport-admin
-  annotations:
-    kubernetes.io/service-account.name: multiport-admin
-type: kubernetes.io/service-account-token
----
-apiVersion: v1
-kind: Secret
-metadata:
   name: multiport
   annotations:
     kubernetes.io/service-account.name: multiport
