+++ acceptance/tests/fixtures/bases/v2-multiport-app/serviceaccount.yaml
@@ -5,8 +5,3 @@ apiVersion: v1
 kind: ServiceAccount
 metadata:
   name: multiport
----
-apiVersion: v1
-kind: ServiceAccount
-metadata:
-  name: multiport-admin
