+++ control-plane/connect-inject/webhook_v2/consul_dataplane_sidecar.go
@@ -12,6 +12,7 @@ import (
 	"github.com/google/shlex"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
+	"k8s.io/apimachinery/pkg/util/intstr"
 	"k8s.io/utils/pointer"
 
 	"github.com/hashicorp/consul-k8s/control-plane/connect-inject/common"
@@ -46,29 +47,29 @@ func (w *MeshWebhook) consulDataplaneSidecar(namespace corev1.Namespace, pod cor
 
 	containerName := sidecarContainer
 
-	//var probe *corev1.Probe
-	//if useProxyHealthCheck(pod) {
-	//	// If using the proxy health check for a service, configure an HTTP handler
-	//	// that queries the '/ready' endpoint of the proxy.
-	//	probe = &corev1.Probe{
-	//		ProbeHandler: corev1.ProbeHandler{
-	//			HTTPGet: &corev1.HTTPGetAction{
-	//				Port: intstr.FromInt(constants.ProxyDefaultHealthPort),
-	//				Path: "/ready",
-	//			},
-	//		},
-	//		InitialDelaySeconds: 1,
-	//	}
-	//} else {
-	//	probe = &corev1.Probe{
-	//		ProbeHandler: corev1.ProbeHandler{
-	//			TCPSocket: &corev1.TCPSocketAction{
-	//				Port: intstr.FromInt(constants.ProxyDefaultInboundPort),
-	//			},
-	//		},
-	//		InitialDelaySeconds: 1,
-	//	}
-	//}
+	var probe *corev1.Probe
+	if useProxyHealthCheck(pod) {
+		// If using the proxy health check for a service, configure an HTTP handler
+		// that queries the '/ready' endpoint of the proxy.
+		probe = &corev1.Probe{
+			ProbeHandler: corev1.ProbeHandler{
+				HTTPGet: &corev1.HTTPGetAction{
+					Port: intstr.FromInt(constants.ProxyDefaultHealthPort),
+					Path: "/ready",
+				},
+			},
+			InitialDelaySeconds: 1,
+		}
+	} else {
+		probe = &corev1.Probe{
+			ProbeHandler: corev1.ProbeHandler{
+				TCPSocket: &corev1.TCPSocketAction{
+					Port: intstr.FromInt(constants.ProxyDefaultInboundPort),
+				},
+			},
+			InitialDelaySeconds: 1,
+		}
+	}
 
 	container := corev1.Container{
 		Name:      containerName,
@@ -107,15 +108,15 @@ func (w *MeshWebhook) consulDataplaneSidecar(namespace corev1.Namespace, pod cor
 				Name:  "DP_PROXY_ID",
 				Value: "$(POD_NAME)",
 			},
-			{
-				Name:  "DP_PROXY_NAMESPACE",
-				Value: "$(POD_NAMESPACE)",
-			},
-			{
-				Name: "DP_PROXY_PARTITION",
-				// TODO: actually use the right partition
-				Value: "default",
-			},
+			//{
+			//	Name:  "DP_PROXY_NAMESPACE",
+			//	Value: "$(POD_NAMESPACE)",
+			//},
+			//{
+			//	Name: "DP_PROXY_PARTITION",
+			//	// TODO: actually use the right partition
+			//	Value: "default",
+			//},
 			{
 				Name:  "DP_CREDENTIAL_LOGIN_META",
 				Value: "pod=$(POD_NAMESPACE)/$(POD_NAME)",
@@ -134,8 +135,12 @@ func (w *MeshWebhook) consulDataplaneSidecar(namespace corev1.Namespace, pod cor
 			},
 		},
 		Args: args,
-		// failing right now with tproxy because expose paths are not implemented
-		//ReadinessProbe: probe,
+	}
+
+	// Omit the readiness probe in transparent proxy mode until expose paths are implemented. Otherwise all probes will fail.
+	// TODO: (v2/nitya) add probes in tproxy mode when expose paths and L7 are supported.
+	if !w.EnableTransparentProxy {
+		container.ReadinessProbe = probe
 	}
 
 	if w.AuthMethod != "" {
